Fix storage and retrieval of Projects
This fix currently works to store Project objects and retreive
them without exposing any of the properties. It uses setProject
as a method to translate stored properties into the variables contained
in the project. A more intuitive solution might be to implement
Projects as a class, then call two different constructors that
take different variables depending on whether it's a new Project
or a storedProject.