Add toDo form
Currently, user is able to add project via form. However,
the toDo form is experiencing an error that wipes the console,
so it's been difficult to troubleshoot. When I simply recreated the
the project form on the 'Add To Do' page, it worked correctly. Greater
attention should be paid to the toDo form to correct the error.